Split post_process.py into separate files
