+++ switch_model/wecc/get_inputs/post_process_steps/graph_config/graph_tech_colors.csv
